% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographyMicrosim.R
\docType{package}
\name{demographyMicrosim}
\alias{demographyMicrosim}
\alias{demographyMicrosim-package}
\title{demographyMicrosim}
\description{
R package for containing demographic examples of static and dynamic microsimulation/microsynthesis
}
\details{
See README.md for detailed information.
}
\section{Functions}{

\code{\link{diversity}}

\code{\link{growth}}

\code{\link{map}}

\code{\link{microsimulate}}

\code{\link{microsynthesise}}

\code{\link{pyramid}}
}

\section{Data}{

\code{\link{msoa}}

\code{\link{sexAgeEth}}

\code{\link{sexAgeYear}}

\code{\link{TowerHamletsFertility}}

\code{\link{TowerHamletsMortality}}
}

