% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diversity}
\alias{diversity}
\title{Calculates a diversity coefficient for a population, grouped by by MSOA}
\usage{
diversity(synpop)
}
\description{
Calculates a diversity coefficient for a population, grouped by by MSOA
}
\examples{
synpop = microsynthesise()
diversity = diversity(synPop)
}
