% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{growth}
\alias{growth}
\title{Calculates growth rate by MSOA between population projections}
\usage{
growth(pop0, pop1)
}
\description{
Calculates growth rate by MSOA between population projections
}
\examples{
basepop = microsynthesise()
projpop = microsimulate(basepop,1)
g = growth(basePop, projpop)
}
