% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{map}
\usage{
map(data, minColour = "#0080FF", maxColour = "#FF8000")
}
\arguments{
\item{data}{a data.table (or data.frame) containing a column "MSOA" containing the MSOA code and
a column "Value" containing numeric data to be plotted}

\item{minColour}{colour for the lowest value of the data}

\item{maxColour}{colour for the lowest value of the data}
}
\description{
Displays a map outlining the MSOAs within Tower Hamlets, filling with a colour derived from the supplied data
}
\examples{
p = microsynthesise()
d = diversity(p)
m = map(d)
}
