% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsimulation.R
\name{microsimulate}
\alias{microsimulate}
\title{microsimulate}
\usage{
microsimulate(synpop, years)
}
\arguments{
\item{synpop}{a dataframe containing the base population to be projected}

\item{years}{a number of years to run the simulation}
}
\value{
a data.table containing the projected population
}
\description{
This function projects an (synthetic) population from aggregate census data. Members of the population are
categorised by age, sex, ethnicity and geographical location.
- births are a single child only (assumed that twins etc are accounted for in the fertility rate)
- newborns are equally likely to be male or female (this may be incorrect esp. for certain ethnicities)
- ethnicity of newborn is the same as mother (in the absence of other information)
- the death rate for age zero is the average of the rate of stillbirths and deaths under one year
- births supersede deaths: thus a person who gives birth and dies in a single year will have a surviving child
}
\examples{
\dontrun{
synpop = microsynthesise()
projpop = microsimulate(synpop, 10)
}
}
