% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsynthesis.R
\name{microsynthesise}
\alias{microsynthesise}
\title{Population microsynthesis}
\usage{
microsynthesise()
}
\description{
This function generates an individual population from aggregate census data. Members of the population are
categorised by age, sex, ethnicity and geographical location.
This file synthesises a population in: MSOA, Sex, Age(single year), Ethnicity
The input data derive from 2011 census tables containing:
  Aggregates of Persons per MSOA, Sex, Age(band) and Ethnicity
  Aggregates of Persons per MSOA, Sex, Age(single year)
}
\examples{
synpop = microsynthesise()
}
