% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{pyramid}
\alias{pyramid}
\title{Produce a pyramid plot for a particular ethnicity}
\usage{
pyramid(eth, population, title)
}
\description{
Produce a pyramid plot for a particular ethnicity
}
\examples{
synpop = pyramid("BAN", synpop11, "Tower Hamlets Bangladeshi population 2011")
}
