% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{pyramid}
\alias{pyramid}
\title{pyramid}
\usage{
pyramid(eth, population, title = "")
}
\arguments{
\item{eth}{a three-letter string identifying the ethnicity}

\item{population}{the overall population}

\item{title}{and optional title for the graph}
}
\description{
Produces a pyramid plot for a particular ethnicity
}
\examples{
p = microsynthesise()
pyramid("BAN", p, "Tower Hamlets Bangladeshi population 2011")
}
